fix(platform): handle listener promises and isolate safeUnlisten import
The typed-lint ratchet rejected the migrated menu handlers in
useRecentFolderEvents and the async settings-deleted handler in
settingsSync: no-floating-promises runs with ignoreVoid=false, so
`void promise` still counts. Each handler now attaches a catch that
logs the failure, and the async atom writer is wrapped in a sync
callback.

useTauriListen also imported `@src/util/platform/tauri`, whose index
runs patchTauriInternals() at import time and re-arms setTimeout
retries for ~1.7s outside Tauri. A test that installs fake timers
mid-chain then observed a phantom pending timer (useEmbeddedWebview
in CI). safeUnlisten moves to a side-effect-free leaf module that the
index re-exports, and a static-import-graph test pins the boundary.
